requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Flask==2.3.3
fonttools==4.42.1
gmplot==1.4.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.24.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
Pillow==10.0.1
plotly==4.0.0
pycparser==2.21
pygame==2.1.3.dev8
pymongo==4.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
retrying==1.3.4
scapy==2.5.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0
